pyproject.toml
[tool.poetry]
name = "gpt-from-scratch-on-pytorch"
version = "0.1.0"
description = ""
authors = ["Hao"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.1"
pyyaml = "^6.0.1"
numpy = "^1.26.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
